refactor: 칵테일 레시피 페이지 컴포넌트 분리
